requirements.txt
pillow~=11.1.0
google-generativeai==0.8.4
